% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal.R
\name{animals}
\alias{animals}
\title{Title}
\usage{
animals(survey_id, project_id, raw = FALSE)
}
\arguments{
\item{animal_id}{}
}
\value{
A data frame or nested list containing information baout animals in the survey
}
\description{
Title
}
