% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report_template}
\alias{create_report_template}
\title{Create the template of a report for a Survey.}
\usage{
create_report_template(
  survey_id,
  project_id,
  abstract_text = "abstract",
  prepared_for_text = "sample",
  background_text = "background"
)
}
\arguments{
\item{project_id}{}
}
\description{
Create the template of a report for a Survey.
}
