% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_all_projects}
\alias{get_all_projects}
\title{Get All Projects available to the current user}
\usage{
get_all_projects()
}
\value{
A dataframe
}
\description{
Get All Projects available to the current user
}
