% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_my_projects}
\alias{get_my_projects}
\title{Get Projects that the current user belongs to}
\usage{
get_my_projects()
}
\value{
A dataframe
}
\description{
Get Projects that the current user belongs to
}
