% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{get_project_details}
\alias{get_project_details}
\title{Get details for a specific Project}
\usage{
get_project_details(project_id)
}
\arguments{
\item{project_id}{Unique ID of a Project for which to get details.}
}
\value{
An object of class Project
}
\description{
Get details for a specific Project
}
