% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_survey_details}
\alias{get_survey_details}
\title{Get details for a specific Survey}
\usage{
get_survey_details(survey_id, project_id)
}
\arguments{
\item{survey_id}{The integer ID of a Survey for which to get details.}

\item{project_id}{The integer ID of a Project from which to get Surveys.}
}
\value{
An object of class Survey
}
\description{
Get details for a specific Survey
}
