% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{get_surveys}
\alias{get_surveys}
\title{Get Surveys in a specific Project}
\usage{
get_surveys(project_id)
}
\arguments{
\item{project_id}{The integer ID of a Project from which to get Surveys.}
}
\value{
A dataframe
}
\description{
Get Surveys in a specific Project
}
