% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_details.numeric}
\alias{survey_details.numeric}
\title{#' @export
survey_details.character <- function(survey_id, project_id, raw = FALSE){
stop("\code{survey_id} must be a positive integer, not a character", call. = FALSE)
}}
\usage{
\method{survey_details}{numeric}(survey_id, project_id, raw = FALSE)
}
\description{
#' @export
survey_details.character <- function(survey_id, project_id, raw = FALSE){
stop("\code{survey_id} must be a positive integer, not a character", call. = FALSE)
}
}
